renaming alembic to fbopenapp.
